adicionado relacao endereco no especialista
