+++ src/main/java/br/com/fiap/Challengesprint3/repository/EstadoRepository.java
@@ -1,9 +1,9 @@
-package br.com.fiap.Challengesprint3.repository;
+// package br.com.fiap.Challengesprint3.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+// import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.com.fiap.Challengesprint3.models.Estado;
+// import br.com.fiap.Challengesprint3.models.Estado;
 
-public interface EstadoRepository extends JpaRepository<Estado, Long>{
+// public interface EstadoRepository extends JpaRepository<Estado, Long>{
     
-}
+// }
