+++ src/main/java/br/com/fiap/Challengesprint3/services/BairroService.java
@@ -1,30 +1,30 @@
-package br.com.fiap.Challengesprint3.services;
+// package br.com.fiap.Challengesprint3.services;
 
-import java.util.List;
-import java.util.Optional;
+// import java.util.List;
+// import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.stereotype.Service;
 
-import br.com.fiap.Challengesprint3.models.Bairro;
-import br.com.fiap.Challengesprint3.repository.BairroRepository;
+// import br.com.fiap.Challengesprint3.models.Bairro;
+// import br.com.fiap.Challengesprint3.repository.BairroRepository;
 
-@Service
-public class BairroService {
+// @Service
+// public class BairroService {
     
-    @Autowired
-    BairroRepository repository;
+//     @Autowired
+//     BairroRepository repository;
 
-    public List<Bairro> listAll() {
-        return repository.findAll();
-    }
+//     public List<Bairro> listAll() {
+//         return repository.findAll();
+//     }
 
-    public Bairro save(Bairro bairro) {
-        return repository.save(bairro);
-    }
+//     public Bairro save(Bairro bairro) {
+//         return repository.save(bairro);
+//     }
 
-    public Optional<Bairro> getById(Long id) {
-        return repository.findById(id);
-    }
+//     public Optional<Bairro> getById(Long id) {
+//         return repository.findById(id);
+//     }
 
-}
+// }
